refactor(agents-api): Rework routers to split routes into individual files
