+++ agents-api/agents_api/routers/sessions/list_sessions.py
@@ -1,0 +1,34 @@
+from fastapi import APIRouter, Depends
+from typing import Annotated
+from pydantic import UUID4
+from agents_api.dependencies.developer_id import get_developer_id
+from agents_api.models.session.list_sessions import list_sessions_query
+from agents_api.autogen.openapi_model import Session
+
+router = APIRouter()
+
+class SessionList(BaseModel):
+    items: list[Session]
+
+@router.get("/sessions", tags=["sessions"])
+async def list_sessions(
+    x_developer_id: Annotated[UUID4, Depends(get_developer_id)],
+    limit: int = 100,
+    offset: int = 0,
+    metadata_filter: str = "{}",
+) -> SessionList:
+    try:
+        metadata_filter = json.loads(metadata_filter)
+    except JSONDecodeError:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="metadata_filter is not a valid JSON",
+        )
+
+    query_results = list_sessions_query(
+        x_developer_id, limit, offset, metadata_filter=metadata_filter
+    )
+
+    return SessionList(
+        items=[Session(**row.to_dict()) for _, row in query_results.iterrows()]
+    )
